+++ packages/core/utils/src/common/index.ts
@@ -92,7 +92,6 @@ export * from "./try-convert-to-boolean"
 export * from "./try-convert-to-number"
 export * from "./unflatten-object-keys"
 export * from "./upper-case-first"
-export * from "./uncountable-words"
 export * from "./validate-handle"
 export * from "./validate-module-name"
 export * from "./wrap-handler"
